+++ test/nester_test.dart
@@ -4,11 +4,14 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:nester/nester.dart';
 
 void main() {
-  test('empty values', () {
+  /// --------------------------------------------------
+  /// LIST tests
+  ///
+  test('list empty values', () {
     bool ok = false;
     try {
       // Empty children
-      const Nester([]);
+      Nester.list([]);
 
       ok = true;
     } on Exception {
@@ -18,10 +21,10 @@ void main() {
     expect(ok, true);
   });
 
-  test('Widgets list', () {
+  test('list', () {
     bool ok = false;
     try {
-      Nester([
+      Nester.list([
         (next) => MaterialApp(
               title: 'Nester library',
               home: next,
@@ -35,6 +38,40 @@ void main() {
               body: next,
             ),
         (next) => Center(child: next),
+        (next) => const Text(
+              'Hello World!',
+              textAlign: TextAlign.center,
+            ),
+      ]);
+
+      ok = true;
+    } on Exception {
+      ok = false;
+    }
+
+    expect(ok, true);
+  });
+
+  /// --------------------------------------------------
+  /// QUEUE tests
+  ///
+  test('queue empty values', () {
+    bool ok = false;
+    try {
+      Nester.queue([
+        (next) => MaterialApp(
+              title: 'Nester library',
+              home: next(),
+            ),
+        (next) => Scaffold(
+              appBar: AppBar(
+                title: const Center(
+                  child: Text('Nester library'),
+                ),
+              ),
+              body: next(),
+            ),
+        (next) => Center(child: next()),
         (next) => const Text(
               'Hello World!',
               textAlign: TextAlign.center,