+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.4
+
+* Deprecated old constructor
+* Added "list" constructor to the list management
+* Added "queue" constructor to the queue management
+* Review the documentation
+
 ## 0.0.3
 
 * Updated description in yaml file